+++ src/user/cs_user_mesh.c
@@ -195,5 +195,23 @@ cs_user_mesh_bad_cells_tag(cs_mesh_t             *mesh,
 }
 
 /*----------------------------------------------------------------------------*/
+/*!
+ * \brief Apply partial modifications to the mesh after the preprocessing
+ *        and initial postprocessing mesh building stage.
+ *
+ * \param[in,out] mesh  pointer to a cs_mesh_t structure
+ * \param[in,out] mesh_quantities pointer to a cs_mesh_quantities_t structure
+*/
+/*----------------------------------------------------------------------------*/
+
+void
+cs_user_mesh_modify_partial(cs_mesh_t             *mesh,
+                            cs_mesh_quantities_t  *mesh_quantities)
+{
+  CS_UNUSED(mesh);
+  CS_UNUSED(mesh_quantities);
+}
+
+/*----------------------------------------------------------------------------*/
 
 END_C_DECLS